Move LazyflowSource to separate file
