+++ volumina/__init__.py
@@ -50,15 +50,5 @@ from . import api
 import volumina.icons_rc
 
 
-def strSlicing(slicing):
-    str = "("
-    for i, s in enumerate(slicing):
-        str += "%d:%d" % (s.start, s.stop)
-        if i != len(slicing) - 1:
-            str += ","
-    str += ")"
-    return str
-
-
 def strQRect(qrect):
     return "(x=%d,y=%d,w=%d,h=%d)" % (qrect.x(), qrect.y(), qrect.width(), qrect.height())
